Update 1048-uk-gemini-encoding-guidance.asciidoc
Some minor corrections:
section numbering leads to duplication of table of contents; I removed the "manual" one - on preview, this loses both, so I'll have to wait to see what happens on publication (to dev)

According to https://github.com/agiorguk/gemini/issues/2, the sample links should already have been to GitHub, although as of 2023-03-08, that is only true of one of each link on the live site. To save migrating the invalid XML samples from the live site (!), I've changed the second of each pair of links to point to GitHub (and removed the 'series' example link).

Need to migrate 1046-xml-order, or remove a lot of links to it (as it is no longer possible to automatically generate it as we could with the old "massive XML -> HTML" process.)

Fixed "end note" links to go directly to the relevant page (i.e. not [4] -> page 1047 "bibliography" under a different name -> an old edition of GEMINI!) - in fact, I removed that link as basically circular!

The source (live) page included a duplicate of "Figure 2" and the following explanatory paragraph; the duplicate figure had been migrated (but not the following paragraph). I've removed it. Similarly for the explanation of Figure 22a. Corrected 2.4.2 reference to "Figure 23" to the correct Figure 25.

How  to avoid URI namespace identifiers being converted to links in ASCII Doc? Fortunately, some of them have landing pages...

Manually numbered the "figures". They (the code fragments) look good in preview, less good in HTML)